Split large serializer file into smaller files
